fix: resolve multiple Excel bugs found in user interview rounds 3-4
- Fix scatter chart --json crash by casting markerSize from byte to int
- Fix shadow=true crash in shape/textbox Add by handling boolean values
- Fix definedNames schema ordering when pivot tables exist
- Fix InlineString cells returning empty text in Get
- Fix sort crash on mixed types by using rank-based comparison
- Fix cell Add ignoring address property (alias for ref)
- Add rowbreak/colbreak Add API for page break management
- Add table totalRow content generation with SUBTOTAL formulas
- Add query("sheet") returning sheet nodes instead of cells
- Add comprehensive tests for all fixes

+++ src/officecli/Core/ChartReader.cs
@@ -142,7 +142,7 @@ internal static partial class ChartHelper
                     seriesNode.Format["marker"] = markerSymbol.InnerText;
                 var markerSize = marker?.GetFirstChild<C.Size>()?.Val;
                 if (markerSize?.HasValue == true)
-                    seriesNode.Format["markerSize"] = markerSize.Value;
+                    seriesNode.Format["markerSize"] = (int)markerSize.Value;
                 node.Children.Add(seriesNode);
             }
             node.ChildCount = seriesList.Count;
